requirements.txt
aioice
attrs
cffi
crcmod
cryptography
pyee
pylibsrtp
pyopenssl
sphinxcontrib-asyncio
pyzmq
numpy
cython
yolo34py-gpu
pandas
glob3
aiohttp
aiortc
opencv-python
